LPS-166857 Moves LayoutStructureUtil to layout-api
